feat: vdf service base class
