+++ src/core/consensus/services/VDFNode.js
@@ -1,13 +1,12 @@
-import ReadyResource from 'ready-resource';
+import { VDFService } from './VDFService.js';
 
-export class VDFNodeService extends ReadyResource {
+export class VDFNode extends VDFService {
     #thread = null;
-    #port = null;
 
     async _open() {
         const { Worker, MessageChannel } = await import('worker_threads');
         const { port1, port2 } = new MessageChannel();
-        this.#port = this.#wrapNodePort(port1);
+        this._setPort(this.#wrapNodePort(port1));
         this.#thread = new Worker(new URL('./vdf-worker.js', import.meta.url), {
             workerData: { port: port2 },
             transferList: [port2],
@@ -16,18 +15,7 @@ export class VDFNodeService extends ReadyResource {
 
     async _close() {
         await this.#thread.terminate();
-        this.#port.close();
-    }
-
-    async calculateVDF(challenge, difficulty, discriminantSizeBits) {
-        await this.#port.write({ challenge, difficulty, discriminantSizeBits });
-        try {
-            const response = await this.#port.read();
-            if (response.error) return null;
-            return response.result;
-        } catch {
-            return null;
-        }
+        await super._close();
     }
 
     #wrapNodePort(port) {
