+++ src/core/consensus/services/VDFService.js
@@ -1,20 +1,14 @@
 import ReadyResource from 'ready-resource';
 
-export class VDFBareService extends ReadyResource {
-    #thread = null;
+export class VDFService extends ReadyResource {
     #port = null;
 
-    async _open() {
-        const { default: Channel } = await import('bare-channel');
-        const channel = new Channel();
-        this.#port = channel.connect();
-        this.#thread = new globalThis.Bare.Thread('./vdf-worker.js', { data: channel.handle });
+    _setPort(port) {
+        this.#port = port;
     }
 
     async _close() {
-        await this.#thread.terminate();
-        await this.#thread.join();
-        await this.#port.close();
+        await this.#port?.close();
     }
 
     async calculateVDF(challenge, difficulty, discriminantSizeBits) {
